Fixed python errors + Added Game Renderer options
* Moved assets to gameConfigs folder

* Moved assets to new folder resources

* Check if is local module resource or absolute

* Load resources module path + Added manifest

* Updated Manifest + Updated TBS games to new resources path

* Cleaned main + Fixed BasicTBS errors with bad configuration

* Added to RTS yamls the GameRenderer section + Updated rts yamls

* Updated test resources paths

* Refactored

+++ core/include/Stratega/GUI/RTSGameRenderer.h
@@ -15,7 +15,7 @@ namespace SGA
 	class RTSGameRenderer : public GameRenderer
 	{
 	public:
-		RTSGameRenderer(SGA::Vector2f& resolution);
+		RTSGameRenderer(SGA::Vector2i& resolution);
 		
 		void init(const GameState& initialState, const GameConfig& gameConfig) override;
 		void update(const GameState& state) override;
